tests/wpt/web-platform-tests/workers/interfaces/WorkerGlobalScope/location/returns-same-object.html

<!--
postMessage(location === location);
/*
-->
<!doctype html>
<title>location === location</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('#');
  worker.onmessage = this.step_func(function(e) {
    assert_true(e.data);
    this.done();
  });
});
</script>
<!--
*/
//-->
